rework scan flow and fix scan history bugs
