+++ web/package.json
@@ -30,6 +30,7 @@
     "react-dom": "18.2.0",
     "react-leaflet": "^4.2.1",
     "react-otp-input": "^3.1.1",
+    "request-ip": "^3.3.0",
     "stripe": "^14.13.0",
     "tailwindcss": "3.3.3",
     "typescript": "5.1.3",
